rag-bot/requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-google-genai==0.0.11
sentence-transformers==2.5.1
chromadb==0.4.24
tenacity==8.2.3
streamlit==1.32.2
python-dotenv==1.0.1
